storybook_crewai/pyproject.toml
[project]
name = "storybook-crewai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.118.0",
    "crewai-tools>=0.43.0",
    "elevenlabs>=1.57.0",
    "google-generativeai>=0.8.5",
    "moviepy>=2.1.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]
